Updated table and debug messages to test

+++ apps/fac/fsw/src/fac_tbldefs.h
@@ -290,7 +290,7 @@ typedef struct
     * @boolean
     * @group FW Attitude Control
     */
-   float FW_W_EN;
+   int32 FW_W_EN;
 
    /**
     * Wheel steering rate proportional gain
